+++ gen/groundobjectsgen.py
@@ -80,6 +80,10 @@ class GroundObjectsGenerator:
                                         vehicle.heading = u.heading
                                         vehicle.player_can_drive = True
                                         vg.add_unit(vehicle)
+
+                                if hasattr(utype, 'eplrs'):
+                                    if utype.eplrs:
+                                        vg.points[0].tasks.append(EPLRS(vg.id))
                             else:
                                 vg = self.m.ship_group(side, g.name, utype, position=g.position,
                                                           heading=g.units[0].heading)